[FLINK-1669] [streaming] [tests] Test for streaming recovery from process failure
